sdk/pyproject.toml
[project]
name = "rhesis-sdk"
version = "0.5.2"
description = "SDK for testing and validating LLM applications"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Engineering Team", email = "engineering@rhesis.ai" },
    { name = "Harry Cruz", email = "harry@rhesis.ai" },
    { name = "Nicolai Bohn", email = "nicolai@rhesis.ai" },
]
keywords = [
    "llm",
    "testing",
    "ai",
    "machine-learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pandas>=2.2.2",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "jinja2>=3.1.6",
    "markitdown[pdf,docx,pptx,xlsx]>=0.1.0",
    "deepeval==3.7.0",
    "tenacity>=8.2.3",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "litellm>=1.76.0",
    "deepteam>=0.2.5",
    "langchain-google-genai>=2.1.12",
    "ragas>=0.3.7",
    "mcp>=1.2.0",
    "websockets>=12.0",
    "ipython>=8.37.0",
    "ipykernel>=7.1.0",
    "opentelemetry-api>=1.32.1",
    "opentelemetry-sdk>=1.32.1",
    "opentelemetry-exporter-otlp-proto-http>=1.32.1",
]

[project.license]
text = "MIT"

[project.urls]
Homepage = "https://rhesis.ai"
Repository = "https://github.com/rhesis-ai/rhesis"
Documentation = "https://rhesis-sdk.readthedocs.io"
"Bug Tracker" = "https://github.com/rhesis-ai/rhesis/issues"


[project.optional-dependencies]
huggingface = [
    "torch>=2.8.0",
    "transformers==4.57.1",
    "accelerate>=1.2.0",
    "bitsandbytes>=0.45.0",
]

# Framework integrations for auto-instrumentation
langchain = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
]

langgraph = [
    "langgraph>=0.0.26",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
]

autogen = [
    "pyautogen>=0.2.0",
]

all-integrations = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.26",
    "pyautogen>=0.2.0",
]

all = [
    "torch>=2.8.0",
    "transformers==4.57.1",
    "accelerate>=1.2.0",
    "bitsandbytes>=0.45.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.26",
    "pyautogen>=0.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "hatch>=1.14.1",
    "pytest>=8.3.4",
    "mypy==1.15.0",
    "types-requests>=2.32.0",
    "pandas-stubs>=2.2.3.250308",
    "types-pytz>=2025.2.0.20250326",
    "types-tqdm>=4.67.0.20241221",
    "sphinx>=8.1.3",
    "sphinx-autodoc-typehints>=3.0.1",
    "sphinx-rtd-theme>=3.0.2",
    "myst-parser>=4.0.0",
    "ruff>=0.11.6",
    "pyright>=1.1.407",
    "psycopg2-binary==2.9.9", # to match the version in the backend
    "pytest-cov>=7.0.0",
]

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/rhesis",
]
include = [
    "src/rhesis/py.typed",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pyright]
typeCheckingMode = "standard"
include = [ "src/rhesis/sdk/metrics", ]

[tool.pytest.ini_options]
pythonpath = [
    "src",
]

# Asyncio configuration
asyncio_default_fixture_loop_scope = "function"

# Filter warnings
filterwarnings = [
    # Ignore Pydantic V2 deprecation warnings from external dependencies
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
]
